# docs/swagger.yaml
basePath: /
definitions:
...
  models.CountProdutsResponse:
    properties:
      count:
        type: integer
      manufacturer:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        description: Код ошибки, например, 400 или 500
        type: integer
      message:
        description: Сообщение об ошибке
        type: string
    type: object
  models.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  models.Product:
    properties:
      category_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      manufacturer:
        type: string
      name:
        type: string
      price:
        type: number
      rating:
        type: number
    type: object
...
  models.ProductResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: Support Team
    url: http://example.com/support
  description: API для интернет-магазина спортивного питания. Содержит функционал
    для управления пользователями, продуктами, заказами и отзывами.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: Sports Nutrition Store API
  version: "1.0"
paths:
...
  /products:
    get:
      consumes:
      - application/json
      description: Получает список продуктов с применением фильтров, сортировки и
        пагинации с тайм-аутом в 2 секунды
      parameters:
      - description: токен
        in: header
        name: Authorization
        type: string
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 10
        description: Количество элементов на странице
        in: query
        name: limit
        type: integer
      - default: id
        description: Поле для сортировки
        in: query
        name: sort
        type: string
      - default: asc
        description: Направление сортировки
        in: query
        name: order
        type: string
      - description: Название продукта
        in: query
        name: name
        type: string
      - description: ID категории
        in: query
        name: category_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешный запрос
          schema:
            $ref: '#/definitions/models.ProductResponse'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Продукты не найдены
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "408":
          description: Тайм-аут запроса
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение списка продуктов с тайм-аутом
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Создает новый продукт с указанными параметрами
      parameters:
      - description: токен
        in: header
        name: Authorization
        type: string
      - description: Данные продукта
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Успешное создание
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создание нового продукта
      tags:
      - products
  /products/{id}:
    delete:
      description: Удаляет продукт по указанному ID
      parameters:
      - description: токен
        in: header
        name: Authorization
        type: string
      - description: ID продукта
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление продукта
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "404":
          description: Продукт не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление продукта
      tags:
      - products
    get:
      description: Получает информацию о продукте по уникальному идентификатору
      parameters:
      - description: токен
        in: header
        name: Authorization
        type: string
      - description: ID продукта
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешный запрос
          schema:
            $ref: '#/definitions/models.Product'
        "404":
          description: Продукт не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение продукта по ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Обновляет данные продукта по указанному ID
      parameters:
      - description: токен
        in: header
        name: Authorization
        type: string
      - description: ID продукта
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные продукта
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Продукт не найден
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновление продукта
      tags:
      - products
...
  /products/count-by-manufacturer:
    get:
      consumes:
      - application/json
      description: Выполняет агрегацию, подсчитывая количество продуктов, сгруппированных
        по производителям.
      parameters:
      - description: токен
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Результаты агрегации с производителями и их количеством
          schema:
            items:
              $ref: '#/definitions/models.CountProdutsResponse'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Подсчет количества продуктов по производителям
      tags:
      - products
  /products/manufacturer:
    put:
      consumes:
      - application/json
      description: Обновляет поле "manufacturer" у всех продуктов в базе данных на
        указанное значение.
      parameters:
      - description: токен
        in: header
        name: Authorization
        type: string
      - description: Новое значение для производителя
        in: query
        name: manufacturer
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера или транзакции
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Массовое обновление производителя продуктов
      tags:
      - products
  /products/price-range:
    get:
      consumes:
      - application/json
      description: Возвращает список продуктов, цены которых находятся в заданном
        диапазоне
      parameters:
      - description: токен
        in: header
        name: Authorization
        type: string
      - description: Минимальная цена
        in: query
        name: minPrice
        required: true
        type: number
      - description: Максимальная цена
        in: query
        name: maxPrice
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: Список продуктов в заданном диапазоне цен
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "400":
          description: Некорректные значения цен
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Продукты не найдены в указанном диапазоне
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение продуктов по диапазону цен
      tags:
      - products
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...